+++ tencentcloud/ess/v20201111/models.go
@@ -4974,6 +4974,11 @@ type ReleasedApprover struct {
 
 	// 要替换的参与人在原合同参与人列表中的签署人编号,通过DescribeFlowInfo 接口获取（即FlowDetailInfos. FlowApproverInfos 结构中的ReceiptId ）
 	RelievedApproverReceiptId *string `json:"RelievedApproverReceiptId,omitempty" name:"RelievedApproverReceiptId"`
+
+	// 指定签署人类型，目前仅支持
+	// ORGANIZATION-企业
+	// ENTERPRISESERVER-企业静默签
+	ApproverType *string `json:"ApproverType,omitempty" name:"ApproverType"`
 }
 
 type RelieveInfo struct {