+++ tencentcloud/partners/v20180321/client.go
@@ -598,6 +598,10 @@ func NewDescribeAgentDealsCacheResponse() (response *DescribeAgentDealsCacheResp
 }
 
 // DescribeAgentDealsCache
+// 已提供新接口并推动切换，改切口目前白名单限制所有访问，申请下线
+//
+// 
+//
 // 【该接口已下线，请使用升级版本DescribeAgentDealsByCache】代理商拉取缓存的全量客户订单
 //
 // 可能返回的错误码:
@@ -610,6 +614,10 @@ func (c *Client) DescribeAgentDealsCache(request *DescribeAgentDealsCacheRequest
 }
 
 // DescribeAgentDealsCache
+// 已提供新接口并推动切换，改切口目前白名单限制所有访问，申请下线
+//
+// 
+//
 // 【该接口已下线，请使用升级版本DescribeAgentDealsByCache】代理商拉取缓存的全量客户订单
 //
 // 可能返回的错误码:
@@ -652,6 +660,10 @@ func NewDescribeAgentPayDealsResponse() (response *DescribeAgentPayDealsResponse
 }
 
 // DescribeAgentPayDeals
+// 已提供新接口并推动切换，改切口目前白名单限制所有访问，申请下线
+//
+// 
+//
 // 【该接口已下线，请切换使用升级版本DescribeAgentPayDealsV2】可以查询代理商代付的所有订单
 //
 // 可能返回的错误码:
@@ -664,6 +676,10 @@ func (c *Client) DescribeAgentPayDeals(request *DescribeAgentPayDealsRequest) (r
 }
 
 // DescribeAgentPayDeals
+// 已提供新接口并推动切换，改切口目前白名单限制所有访问，申请下线
+//
+// 
+//
 // 【该接口已下线，请切换使用升级版本DescribeAgentPayDealsV2】可以查询代理商代付的所有订单
 //
 // 可能返回的错误码:
@@ -760,6 +776,10 @@ func NewDescribeAgentSelfPayDealsResponse() (response *DescribeAgentSelfPayDeals
 }
 
 // DescribeAgentSelfPayDeals
+// 已提供新接口并推动切换，改切口目前白名单限制所有访问，申请下线
+//
+// 
+//
 // 【该接口已下线，请切换使用升级版本DescribeAgentSelfPayDealsV2】可以查询代理商下指定客户的自付订单
 //
 // 可能返回的错误码:
@@ -772,6 +792,10 @@ func (c *Client) DescribeAgentSelfPayDeals(request *DescribeAgentSelfPayDealsReq
 }
 
 // DescribeAgentSelfPayDeals
+// 已提供新接口并推动切换，改切口目前白名单限制所有访问，申请下线
+//
+// 
+//
 // 【该接口已下线，请切换使用升级版本DescribeAgentSelfPayDealsV2】可以查询代理商下指定客户的自付订单
 //
 // 可能返回的错误码: